requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
python-dotenv==1.0.0
openai>=1.3.0
google-search-results==2.4.2
gunicorn==21.2.0